requirements.txt
flask>=3.0.0
gunicorn>=21.2.0
bcrypt>=4.1.0
python-dotenv>=1.0.0
flask-sqlalchemy>=3.1.0
flask-migrate>=4.0.0
psycopg2-binary>=2.9.9
flask-limiter>=3.5.0
cryptography>=42.0.0
